+++ src/sys/fatfs.c
@@ -10,6 +10,11 @@ void dsy_fatfs_init(void)
     retSD = FATFS_LinkDriver(&SD_Driver, SDPath);
 }
 
+void dsy_fatfs_deinit(void)
+{
+    retSD = FATFS_UnLinkDriver(SDPath);
+}
+
 DWORD get_fattime(void)
 {
     return 0;